requirements.txt
google-adk==0.5.0
google-cloud-aiplatform[adk,agent_engines]==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
streamlit
requests
sseclient-py
python-dotenv
